requirements.txt
click==8.1.3
humanize==4.6.0
idna==3.4
keyboard==0.13.5
PyGithub==1.58.1
python-decouple==3.7
retrying==1.3.4
rich==13.3.1
